# swagger-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/suppliers": {
      "get": {
        "operationId": "SupplierController_get",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Página atual",
            "schema": {
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "description": "Itens por página",
            "schema": {
              "example": 10,
              "type": "number"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Busca por nome ou fantasia",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "situation",
            "required": false,
            "in": "query",
            "description": "Situação do fornecedor",
            "schema": {
              "type": "string",
              "enum": [
                "A",
                "I"
              ]
            }
          },
          {
            "name": "storeId",
            "required": false,
            "in": "query",
            "description": "ID da loja",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de fornecedores",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SupplierResponseDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Listar fornecedores",
        "tags": [
          "Supplier"
        ]
      }
    },
    "/api/v1/suppliers/{id}": {
      "get": {
        "operationId": "SupplierController_getById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID do fornecedor (CRE_CODIGO)",
            "schema": {
              "example": 1,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Detalhes do fornecedor",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SupplierDetailResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Obter detalhe do fornecedor",
        "tags": [
          "Supplier"
        ]
      }
    },
...
    "/status/my-connection": {
      "get": {
        "operationId": "StatusController_getMyConnection",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Verificação em tempo real da conexão com API, Postgres e Firebird.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MyConnectionStatusResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Verificar status da conexão do tenant (Usuário)",
        "tags": [
          "Status"
        ]
      }
    },
...
  },
  "info": {
    "title": "InfoBrasil API",
    "description": "API de integração do sistema Retaguarda (by InfoBrasil) com outros sistemas, como ERPs, e-commerce, etc. <br><br>\n      Se deseja solicitar um acesso, preencha o formulário em: <a href='https://info-api.infobrasilsistemas.com.br/integration/form'>Solicitar Acesso</a>\n      <br><br>\n      **O sistema conta com duas formas de autenticação:**<br> \n      - Basic Auth para login e obtenção de token JWT<br>\n      - JWT para acesso aos endpoints protegidos.\n      <br><br>\n      Para verificar a disponibilidade da API acesse nossa <a href='/status'>Página de Status</a>\n      <br><br>\n      A API é organizada em módulos, cada um responsável por uma área específica do sistema, como produtos, clientes, vendas, etc.",
    "version": "1.15.1",
    "contact": {
      "name": "InfoBrasil Sistemas",
      "url": "https://www.infobrasilsistemas.com.br",
      "email": "suporte@infobrasilsistemas.com.br"
    }
  },
...
  "servers": [
    {
      "url": "http://localhost:3336",
      "description": "Servidor local para desenvolvimento"
    },
    {
      "url": "https://info-api.infobrasilsistemas.com.br",
      "description": "Servidor de produção"
    }
  ],
  "components": {
    "securitySchemes": {
      "basic": {
        "type": "http",
        "scheme": "basic"
      },
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "SupplierResponseDto": {
        "type": "object",
        "properties": {
          "CRE_CODIGO": {
            "type": "number",
            "example": 1,
            "description": "Código do fornecedor"
          },
          "CRE_NOME": {
            "type": "string",
            "example": "Fornecedor Exemplo LTDA",
            "description": "Razão social ou nome"
          },
          "CRE_FANTASIA": {
            "type": "string",
            "example": "Fornecedor Exemplo",
            "description": "Nome fantasia"
          },
          "CRE_ENDERECO": {
            "type": "string",
            "example": "Av. Brasil, 1500",
            "description": "Endereço"
          },
          "CRE_BAIRRO": {
            "type": "string",
            "example": "Jardim Paulista",
            "description": "Bairro"
          },
          "CRE_UF": {
            "type": "string",
            "example": "SP",
            "description": "UF"
          },
          "CRE_CONTATO": {
            "type": "string",
            "example": "Carlos",
            "description": "Contato principal"
          },
          "CRE_CEP": {
            "type": "string",
            "example": "01431-000",
            "description": "CEP"
          },
          "CRE_CNPJ": {
            "type": "string",
            "example": "12.345.678/0001-90",
            "description": "CNPJ"
          },
          "CRE_CGF": {
            "type": "string",
            "example": "123456789",
            "description": "Inscrição Estadual (CGF)"
          },
          "CRE_FONE": {
            "type": "string",
            "example": "1133334444",
            "description": "Telefone"
          },
          "CRE_CELULAR": {
            "type": "string",
            "example": "11999998888",
            "description": "Celular"
          },
          "CRE_FAX": {
            "type": "string",
            "example": "1133334445",
            "description": "Fax"
          },
          "CRE_ZEROOITO": {
            "type": "string",
            "example": "0800-123-456",
            "description": "Telefone 0800"
          },
          "CRE_EMAIL": {
            "type": "string",
            "example": "contato@fornecedor.com",
            "description": "E-mail"
          },
          "CRE_SITE": {
            "type": "string",
            "example": "www.fornecedor.com.br",
            "description": "Site"
          },
          "CRE_REPRESENTANTE": {
            "type": "string",
            "example": "Representante Zé",
            "description": "Nome do representante"
          },
          "CRE_FONEREP": {
            "type": "string",
            "example": "1133334444",
            "description": "Telefone do representante"
          },
          "CRE_CELULARREP": {
            "type": "string",
            "example": "11988887777",
            "description": "Celular do representante"
          },
          "CRE_EMAILREP": {
            "type": "string",
            "example": "rep@fornecedor.com",
            "description": "E-mail do representante"
          },
          "CRE_SITUACAO": {
            "type": "string",
            "example": "A",
            "description": "Situação (A = Ativo, I = Inativo)"
          },
          "LOJ_CODIGO": {
            "type": "number",
            "example": 1,
            "description": "Código da loja"
          }
        },
        "required": [
          "CRE_CODIGO",
          "CRE_NOME",
          "CRE_FANTASIA",
          "CRE_ENDERECO",
          "CRE_BAIRRO",
          "CRE_UF",
          "CRE_CONTATO",
          "CRE_CEP",
          "CRE_CNPJ",
          "CRE_CGF",
          "CRE_FONE",
          "CRE_CELULAR",
          "CRE_FAX",
          "CRE_ZEROOITO",
          "CRE_EMAIL",
          "CRE_SITE",
          "CRE_REPRESENTANTE",
          "CRE_FONEREP",
          "CRE_CELULARREP",
          "CRE_EMAILREP",
          "CRE_SITUACAO",
          "LOJ_CODIGO"
        ]
      },
      "SupplierDetailResponseDto": {
        "type": "object",
        "properties": {
          "CRE_CODIGO": {
            "type": "number",
            "example": 1,
            "description": "Código do fornecedor"
          },
          "CRE_NOME": {
            "type": "string",
            "example": "Fornecedor Exemplo LTDA",
            "description": "Razão social ou nome"
          },
          "CRE_FANTASIA": {
            "type": "string",
            "example": "Fornecedor Exemplo",
            "description": "Nome fantasia"
          },
          "CRE_ENDERECO": {
            "type": "string",
            "example": "Av. Brasil, 1500",
            "description": "Endereço"
          },
          "CRE_BAIRRO": {
            "type": "string",
            "example": "Jardim Paulista",
            "description": "Bairro"
          },
          "CRE_UF": {
            "type": "string",
            "example": "SP",
            "description": "UF"
          },
          "CRE_CONTATO": {
            "type": "string",
            "example": "Carlos",
            "description": "Contato principal"
          },
          "CRE_CEP": {
            "type": "string",
            "example": "01431-000",
            "description": "CEP"
          },
          "CRE_CNPJ": {
            "type": "string",
            "example": "12.345.678/0001-90",
            "description": "CNPJ"
          },
          "CRE_CGF": {
            "type": "string",
            "example": "123456789",
            "description": "Inscrição Estadual (CGF)"
          },
          "CRE_FONE": {
            "type": "string",
            "example": "1133334444",
            "description": "Telefone"
          },
          "CRE_CELULAR": {
            "type": "string",
            "example": "11999998888",
            "description": "Celular"
          },
          "CRE_FAX": {
            "type": "string",
            "example": "1133334445",
            "description": "Fax"
          },
          "CRE_ZEROOITO": {
            "type": "string",
            "example": "0800-123-456",
            "description": "Telefone 0800"
          },
          "CRE_EMAIL": {
            "type": "string",
            "example": "contato@fornecedor.com",
            "description": "E-mail"
          },
          "CRE_SITE": {
            "type": "string",
            "example": "www.fornecedor.com.br",
            "description": "Site"
          },
          "CRE_REPRESENTANTE": {
            "type": "string",
            "example": "Representante Zé",
            "description": "Nome do representante"
          },
          "CRE_FONEREP": {
            "type": "string",
            "example": "1133334444",
            "description": "Telefone do representante"
          },
          "CRE_CELULARREP": {
            "type": "string",
            "example": "11988887777",
            "description": "Celular do representante"
          },
          "CRE_EMAILREP": {
            "type": "string",
            "example": "rep@fornecedor.com",
            "description": "E-mail do representante"
          },
          "CRE_SITUACAO": {
            "type": "string",
            "example": "A",
            "description": "Situação (A = Ativo, I = Inativo)"
          },
          "LOJ_CODIGO": {
            "type": "number",
            "example": 1,
            "description": "Código da loja"
          }
        },
        "required": [
          "CRE_CODIGO",
          "CRE_NOME",
          "CRE_FANTASIA",
          "CRE_ENDERECO",
          "CRE_BAIRRO",
          "CRE_UF",
          "CRE_CONTATO",
          "CRE_CEP",
          "CRE_CNPJ",
          "CRE_CGF",
          "CRE_FONE",
          "CRE_CELULAR",
          "CRE_FAX",
          "CRE_ZEROOITO",
          "CRE_EMAIL",
          "CRE_SITE",
          "CRE_REPRESENTANTE",
          "CRE_FONEREP",
          "CRE_CELULARREP",
          "CRE_EMAILREP",
          "CRE_SITUACAO",
          "LOJ_CODIGO"
        ]
      },
...
      "MyConnectionStatusResponseDto": {
        "type": "object",
        "properties": {
          "api": {
            "type": "string",
            "description": "Status da API (UP/DOWN)",
            "example": "UP"
          },
          "apiLatency": {
            "type": "number",
            "description": "Latência da API em milissegundos",
            "example": 5
          },
          "postgres": {
            "type": "string",
            "description": "Status do Banco de Dados Postgres (UP/DOWN)",
            "example": "UP"
          },
          "postgresLatency": {
            "type": "number",
            "description": "Latência do Banco de Dados Postgres em milissegundos",
            "example": 2
          },
          "tenant": {
            "type": "string",
            "description": "Status do Banco de Dados Firebird (Tenant) (UP/DOWN)",
            "example": "UP"
          },
          "tenantLatency": {
            "type": "number",
            "description": "Latência do Banco de Dados Firebird (Tenant) em milissegundos",
            "example": 45
          }
        },
        "required": [
          "api",
          "apiLatency",
          "postgres",
          "postgresLatency",
          "tenant",
          "tenantLatency"
        ]
      }
    }
  }
}